// frontend/app/api/generated/model/getDiary200.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Diary API
 * Backend API for authentication and diary entries.
 * OpenAPI spec version: 1.0.0
 */

export type GetDiary200 = {
  id: string;
  title: string;
  content: string;
  /** ISO 8601 date-time string */
  createdAt: string;
  /** ISO 8601 date-time string */
  updatedAt: string;
};