// frontend/app/api/generated/model/listDiaries200DiariesItem.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Diary API
 * Backend API for authentication and diary entries.
 * OpenAPI spec version: 1.0.0
 */

export type ListDiaries200DiariesItem = {
  id: string;
  title: string;
  contentPreview: string;
  /** ISO 8601 date-time string */
  createdAt: string;
  /** ISO 8601 date-time string */
  updatedAt: string;
};